moved actions and reducers into seperate files
