+++ src/reducers/rootReducer.js
@@ -1,13 +1,23 @@
-import {combineReducers} from 'redux';
-import {logInReducer} from './logInReducer';
-import {registerReducer} from './registerReducer';
-import {groupReducer} from './groupReducer';
+import { combineReducers } from 'redux';
+import { logInReducer } from './logInReducer';
+import { registerReducer } from './registerReducer';
+import { groupReducer } from './groupReducer';
+import { groupPostReducer } from './groupPostReducer';
+import { groupMarkReducer } from './groupMarkReducer';
+import { groupMarkCategoryReducer } from './groupCustomMarkCategoryReducer';
+import { groupEventReducer } from './groupEventReducer';
+import { groupChatRoomReducer } from './groupChatRoomReducer';
 // for future form implementation
 // import {reducer as formReducer} from 'redux-form';
 
 export default combineReducers({
   logInReducer,
   registerReducer,
-  groupReducer
+  groupReducer,
+  groupPostReducer,
+  groupMarkReducer,
+  groupMarkCategoryReducer,
+  groupEventReducer,
+  groupChatRoomReducer,
   //  form: formReducer,
 });
